[MOVE] commandsのパッケージ構造を大幅に変更する
コードロジックへの変更はありません
