+++ src/main/java/com/github/unchama/seichiassist/commands/contextual/builder/ArgumentParserScope.kt
@@ -1,10 +1,10 @@
-package com.github.unchama.seichiassist.commands.abstract
+package com.github.unchama.seichiassist.commands.contextual.builder
 
 import arrow.core.Left
 import arrow.core.None
 import arrow.core.Right
 import arrow.core.Some
-import com.github.unchama.seichiassist.commands.abstract.response.asResponseToSender
+import com.github.unchama.seichiassist.commands.contextual.builder.response.asResponseToSender
 
 object ArgumentParserScope {
