+++ src/main/java/com/github/unchama/seichiassist/commands/contextual/builder/TypeAliases.kt
@@ -1,9 +1,12 @@
-package com.github.unchama.seichiassist.commands.abstract
+package com.github.unchama.seichiassist.commands.contextual.builder
 
 import arrow.core.Either
 import arrow.core.Option
 import arrow.effects.IO
-import com.github.unchama.seichiassist.commands.abstract.response.ResponseToSender
+import com.github.unchama.seichiassist.commands.contextual.ParsedArgCommandContext
+import com.github.unchama.seichiassist.commands.contextual.PartiallyParsedArgs
+import com.github.unchama.seichiassist.commands.contextual.RawCommandContext
+import com.github.unchama.seichiassist.commands.contextual.builder.response.ResponseToSender
 
 typealias CommandResponse = Option<ResponseToSender>
